+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Replace custom implementation of `Math.countlz` with the new engine `bit32.countlz` function
 * Fix `indexOf` to be accessible from `String`
 * Refactor structure to a rotriever workspace
+* Fix Error stacktrace to include error name and message
 
 ## 1.0.0